requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
tortoise-orm[asyncpg]==0.20.0
aerich==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
replicate==0.22.0
boto3==1.34.0
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.1.0
python-multipart==0.0.6